Improved code claridy and increased comedic value
I believe the new when the based system is the future of programming, its significantly easier to follow. Will be seeing much use in my future projects!
Using the when the system I was able to make more references to comedy and in the when the infrastructure I was able to reference Allah basmala brothers.